+++ NEWS.md
@@ -1,3 +1,8 @@
+# ppmf 0.0.3
+* Uses GitHub release links for compressed data
+* Add get_ppmf_links() to get the links for data
+* Basic pkgdown site
+
 # ppmf 0.0.2
 * Added a `NEWS.md` file to track changes to the package.
 * Added function breakdown_geoid() to convert GEOIDs into component parts